packetdrill: run_all: support timeout with --serialized
Before, when using the --serialized argument, the script was simply
waiting for the packetdrill command to finish but without any timeout.

Now that the limitation of the parallel tasks has been put in place, it
is possible to easily support this timeout feature.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
(cherry picked from commit 9f79485415fea0180da07fe8c86f0189c07d6190)
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>