mptcp: tfo: cookie req: remove close delay
This test is unstable for a bit of time. When we get an error, it looks
like it is because the kernel retransmit the DATA_FIN, and the packet at
line 23 is then not the expected one:

  server-TCP_FASTOPEN-cookie-req.pkt:23: error handling packet: live packet field ipv4_total_length: expected: 48 (0x30) vs actual: 64 (0x40)
  script packet:  0.634961 . 1:1(0) ack 1 <dss dack4 3007449510 flags: A>
  actual packet:  0.629870 . 1:1(0) ack 1 win 256 <dss dack4 3007449509 dsn8 17244274742396137526 ssn 0 dll 1 no_checksum flags: MmAF,nop,nop>

The retransmission might be due to the fact the packets are injected
with a delay, causing the RTO to fire, and resulting in a
retransmission. Removing the delay should then help stabilising this
test.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>
(cherry picked from commit c44c7bf39e1c309a0e17f6b2d714c3f2596e5277)
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>