mptcp: more tolerant on RTX timeouts
When running the different tests in a loop, tests looking at
retransmissions can fail because of a timing error, e.g.:

  3rdack_rtx.pkt:34: error handling packet: timing error: expected outbound packet at 1.460229 sec but happened at 1.623767 sec; tolerance 0.100000 sec
  dss_fin_retrans_established.pkt:22: error handling packet: timing error: expected outbound packet at 0.820319 sec but happened at 0.920328 sec; tolerance 0.100000 sec

The goal of the test is to check that the retransmissions happen, the
precision of the time is not that important as it might be delayed due
to various factors. As long as they are fired approximatively when they
are expected is fine. Better to increase the tolerance just for these
packets, not all.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>
(cherry picked from commit d201c78de0e3a16b7396778c71c8f0667bc3ab5b)
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>