mptcp: multi-ep: support v6 subflow endpoints
Compared to v4, with v6, the IP pref is in reverse order of their
definition.

A workaround is then to remove and re-add the old addresses in the
reverse order. If we don't do that, the kernel will initiate the
connection using the address that has been added last, so not 'caddr1'
like it is the case with v4. We need a consistent behaviour, not to have
to duplicate tests for v4 and v6.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>
(cherry picked from commit 20e25564cdf3d97ef3964d8d94b608a1f9309105)
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>